Fix dead lock in Tethering state machine
Tethering is animated by two state machines. If an error occurs,
one of the state machine notifies the other and definitely stays
in the error state while the other retries the tether operation
after having treated the error.

This leads in a permanent loop where one state machine will retry
indefinitely while the other is blocked in error. In such situation
the user sees the hotspot looping ON / OFF until device reboot.

This patch allows to clear the error of the blocked state machine
from the other one and thus recover a normal cycle giving a chance
to use tether feature without rebooting the device.

Change-Id: I6ec3e3affc3c9929fcb06f2d4a1b10c33bda4b8a
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>