Fix softboot when user starts a call when the screen is off and the smart cover is closed
This commits reverts

    policy: fix device sleeping with lid with incoming call

    commit af27fcc05f7119ed908124cf71ddcb981bfe1228.

    and

    policy: respect FLAG_TURN_SCREEN_ON flag with device flip covers

    commit 8b5e80b968eb2f3f91ad90f74d1eeae7d51289f6.

This code was working without issue on lp/cm-12.1, but on MM cause a race condition when the user tries to start a call and the smart cover is closed

Reproducible: always

Steps to reproduce the issue: Simply start a call with the cover closed, sometimes the first call works, but the second call will always stuck the device and cause a softboot.

Tried with a carkit, with adb and airdroid, and we get the same result

@smeroni68 confirmed that even opt-cm-13.0 and cm-13.0 has this issue

Big big thanks to @smeroni68 and @Alucard24 for their help! :-)