Don't take screenshots for tasks that are being cleared
First of all, it's unnecessary. The task is about to be populated with
a new activity, and all the old ones thrown away.

But we also saw some cases where an app would be blocked for 300ms when
calling startActivity() and resetting the task inside onCreate(). This
was because of a retry loop in the screenshot code - the activity being
paused had not drawn anything, and was blocked on the startActivity()
call, so the screenshot failed.

Change-Id: I1bfcea9c1cd3ad9cde153f2064d197551fc19b83