Request focus of status bar again when collapsing is canceled
When the panel collapsing with animateCollapsePanels(),
it releases focus of status bar immediately.
But if collapsing is cancelled by user,
the panel remains out of focus and <Back key> acts with
the background activity, not the expanded panel.

Also, when animateCollapsePanels() is called by <Back key> or etc,
while user is expanding the panel by swiping,
it releases focus of status bar immediately again, too.

So, fix these bugs by re-requesting focus when collapsing is cancelled,
and not releasing focus when the panel is not collapsing.

Also, remove unused variables of PhoneStatusBar.

Change-Id: I16ee9b45ebd8a3624d264d254e0e5da4a55e52f5