bootanimation: fix check for system bootanimation
We already check for the system bootanimation the proper way (using the
getAnimationFileName function), so we shouldn't have re-added the clause to
check it "the old way".

As it was, if there is a system bootanimation on the traditional path, an
alternative system bootanimation provided by getAnimationFileName would not be
used.

Change-Id: I73178d6b819c40e120c42fe10cb20afe171b8d4e
Bug-Id: PAELLA-112

Conflicts:
	cmds/bootanimation/BootAnimation.cpp